requirements.txt
fastapi
uvicorn[standard]
twilio
httpx
pydub
python-multipart
google-genai

